gestion-analyse/requirements.txt
absl-py==2.1.0
accelerate==0.27.0
aiohttp==3.9.3
aiosignal==1.3.1
asgiref @ file:///croot/asgiref_1724072470684/work
async-timeout==4.0.3
attributedict==0.3.0
autoawq==0.1.8
blessings==1.7
cachetools==5.3.2
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
colour-runner==0.1.1
coverage==7.4.1
DataProperty==1.0.1
datasets==2.17.0
deepdiff==6.7.1
dill==0.3.8
distlib==0.3.8
dj-database-url==2.3.0
Django @ file:///croot/django_1724964382619/work
djangorestframework==3.15.2
docker==7.1.0
environs==11.0.0
evaluate==0.4.1
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
greenlet==3.1.1
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.10
inspecta==0.1.3
Jinja2==3.1.3
joblib==1.3.2
jsonlines==4.0.0
lm_eval==0.4.1
lxml==5.1.0
marshmallow==3.22.0
mbstrdecoder==1.1.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
nltk==3.8.1
numexpr==2.9.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
ordered-set==4.1.0
packaging==23.2
pandas==2.2.0
pathvalidate==3.2.0
peft==0.8.2
pillow==10.4.0
platformdirs==4.2.0
pluggy==1.4.0
portalocker==2.8.2
psutil==5.9.8
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2-binary==2.9.10
pyarrow==15.0.0
pyarrow-hotfix==0.6
pybind11==2.11.1
Pygments==2.17.2
pyproject-api==1.6.1
pytablewriter==1.2.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
regex==2023.12.25
requests==2.31.0
responses==0.18.0
rootpath==0.1.1
rouge-score==0.1.2
sacrebleu==2.4.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
sentencepiece==0.1.99
SQLAlchemy==2.0.36
sqlitedict==2.1.0
sqlparse @ file:///croot/sqlparse_1690901927333/work
sympy==1.12
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.4
termcolor==2.4.0
testcontainers==4.8.2
threadpoolctl==3.2.0
tokenizers==0.15.1
toml==0.10.2
tomli==2.0.1
torch==2.2.0
torchvision==0.17.0
tox==4.12.1
tqdm==4.66.1
tqdm-multiprocess==0.0.11
transformers==4.37.2
triton==2.2.0
typepy==1.3.2
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.3
virtualenv==20.25.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zstandard==0.22.0
